Moved all cerebro-related plugins in one directory, added "quit" plugin
