+++ app/main/plugins/core/cerebro/icon.png
