+++ app/main/plugins/core/cerebro/index.js
@@ -1,0 +1,8 @@
+import autocomplete from './autocomplete'
+import quit from './quit'
+import plugins from './plugins'
+import settings from './settings'
+
+module.exports = {
+  autocomplete, quit, plugins, settings
+}
