+++ app/main/plugins/core/cerebro/plugins/Preview/ActionButton.js
