+++ app/main/plugins/core/cerebro/plugins/Preview/index.js
