+++ app/main/plugins/core/cerebro/plugins/Preview/styles.css
