+++ app/main/plugins/core/cerebro/plugins/getAvailablePlugins.js
