+++ app/main/plugins/core/cerebro/plugins/getInstalledPlugins.js
