+++ app/main/plugins/core/cerebro/plugins/getReadme.js
