+++ app/main/plugins/core/cerebro/plugins/index.js
@@ -3,7 +3,7 @@ import Preview from './Preview'
 import { search, memoize } from 'cerebro-tools'
 import availablePlugins from './getAvailablePlugins'
 import installedPlugins from './getInstalledPlugins'
-import icon from './icon.png'
+import icon from '../icon.png'
 
 const getAvailablePlugins = memoize(availablePlugins)
 const getInstalledPlugins = memoize(() => (
