+++ app/main/plugins/core/cerebro/settings/Settings/Hotkey/index.js
