+++ app/main/plugins/core/cerebro/settings/Settings/Hotkey/styles.css
