+++ app/main/plugins/core/cerebro/settings/Settings/index.js
